+++ packages/ActionBar/src/components/filter.js
@@ -1,0 +1,13 @@
+export const Filter = () => {
+  console.error(
+    "The <Filter /> component in @paprika/action-bar has been deprecated. Please use the package @paprika/filter directly."
+  );
+
+  return null;
+};
+
+export const useFilter = () => {
+  console.error(
+    "The hook useFilter and the <Filter /> component in @paprika/action-bar are deprecated. Please use the package @paprika/filter directly."
+  );
+};