+++ packages/ActionBar/src/index.js
@@ -3,4 +3,5 @@ export { default as ColumnsArrangement } from "./components/ColumnsArrangement";
 export { default as SearchInput } from "./components/SearchInput";
 export { default as useSort } from "./hooks/useSort";
 export { default as useColumnsArrangement } from "./hooks/useColumnsArrangement";
+export { Filter, useFilter } from "./components/filter";
 export { default } from "./ActionBar";