+++ packages/ActionBar/CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
-- Removed <Filters> component from this package. Please use `@paprika/filters` instead.
+- Removed <Filter> component and `useFilter` from this package. Please use `@paprika/filter` instead.
 
 ## [0.2.0] - 2020-06-08